Moved Search bar from explore page to its own component
