+++ client/pages/explore.js
@@ -3,6 +3,7 @@ import axios from "axios";
 import { CardsetView } from "@/components/DetailedCardsetView";
 import { ExploreCard } from '@/components/Cards/ExploreCard';
 import Navbar from '@/components/Navbar/Navbar';
+import { SearchBar } from "@/components/Explore/SearchBar";
 
 const Explore = () => {
     const [cardsets, setCardsets] = useState([]);
@@ -16,64 +17,8 @@ const Explore = () => {
         fetchCardsets();
     }, []);
 
-    useEffect(() => {
-        filterCardsets();
-    }, [searchInput]);
-
-    const onSearchChange = (e) =>{
-        e.preventDefault();
-        setSearchInput(e.target.value);
-    }
-
-    const filterCardsets = () => {
-        const searchLower = searchInput.toLowerCase();
-        setFilteredCardsets([...cardsets.filter(cardset => 
-            cardset.title.toLowerCase().includes(searchLower) || 
-            cardset.subject.toLowerCase().includes(searchLower) ||
-            (cardset.user?.username && cardset.user.username.toLowerCase().includes(searchLower))
-        )]);
-    };
-    
-
-    const onSortChangeClicked = (e, sortBy) =>{
-        e.preventDefault();
-        let sortedCardsets;
-        switch(sortBy){
-            case 'flashcardCount':
-                sortedCardsets = sortByFlashcards();
-                break;
-            case 'creationNewest':
-                sortedCardsets = sortByCreation();
-                sortedCardsets.reverse();
-                break;
-            case 'creationOldest':
-                sortedCardsets = sortByCreation();
-                break;
-            case 'alphabeticalOrder':
-                sortedCardsets = sortByAlphabet();
-                break;
-            default: 
-                sortedCardsets = cardsets;
-                break;   
-        }
-        setCardsets([...sortedCardsets]);
-        filterCardsets();
-
-    }
-
-    const sortByFlashcards = () => {
-        setSortingBy('Flashcard Count');
-        return cardsets.sort((a,b) => b.flashcardCount - a.flashcardCount);
-    }
-
-    const sortByCreation = () => {
-        setSortingBy('Creation Date');
-        return cardsets.sort((a,b) => a.createdAt.localeCompare(b.createdAt));
-    }
-
-    const sortByAlphabet = () => {
-        setSortingBy("Alphabetically");
-        return cardsets.sort((a,b) => a.title.localeCompare(b.title));
+    const onSearchUpdate = (sortedSets) => {
+        setFilteredCardsets(sortedSets);
     }
 
     const fetchCardsets = async () => {
@@ -105,22 +50,7 @@ const Explore = () => {
             <Navbar/>
             <div className="container mt-5">
                 <h1 className="mb-4">Explore Cardsets</h1>
-            <div className='d-flex mb-5'>
-                <form className="form-inline" onSubmit={((e) => e.preventDefault())}>
-                    <input className="form-control mr-sm-2" type="search" placeholder="Search" aria-label="Search" onInput={(e) => onSearchChange(e)}/>
-                </form>
-                <div className="dropdown">
-                    <button className="btn btn-secondary dropdown-toggle" type="button" data-bs-toggle="dropdown" aria-expanded="false">
-                        {sortingBy.length > 0 ? sortingBy : "Sort By..."}
-                    </button>
-                    <ul className="dropdown-menu">
-                        <li><a className="dropdown-item" onClick={(e) => onSortChangeClicked(e, 'flashcardCount')}>Flashcard count</a></li>
-                        <li><a className="dropdown-item" onClick={(e) => onSortChangeClicked(e, 'creationNewest')}>Newest first</a></li>
-                        <li><a className="dropdown-item" onClick={(e) => onSortChangeClicked(e, 'creationOldest')}>Oldest first</a></li>
-                        <li><a className="dropdown-item" onClick={(e) => onSortChangeClicked(e, 'alphabeticalOrder')}>Alphabetical order</a></li>
-                    </ul>
-                </div>
-            </div>
+                <SearchBar cardsets={cardsets} onSearchUpdate={onSearchUpdate}/>
 
             <div className="row">
                 { filteredCardsets.length == 0 && searchInput.length > 0 && <div>No cardsets matching this term</div> }
